changed to postgresql

+++ scorecard-performance-tracker/Dtos/ScoreResponseDto.cs
@@ -7,11 +7,11 @@ namespace scorecard_performance_tracker.DTOs
         public string Id { get; set; }
         public string UserId { get; set; }
         public string DevSquad { get; set; }
-        public short AgileScore { get; set; }
-        public short AlgorithmScore { get; set; }
-        public short AssesmentScore { get; set; }
-        public short WeeklyTaskScore { get; set; }
-        public short CumulativeScore { get; set; }
+        public double AgileScore { get; set; }
+        public double AlgorithmScore { get; set; }
+        public double AssesmentScore { get; set; }
+        public double WeeklyTaskScore { get; set; }
+        public double CumulativeScore { get; set; }
         public DateTime CreationTime { get; set; }
         public string CreatedBy { get; set; }
         public DateTime UpdatedOn { get; set; }
